Server Tests - Refactor Server Test To Reduce Deplication
