+++ server/__tests__/api_common.test.js
@@ -1,3 +1,8 @@
+const request = require('supertest')
+const connectDB = require('../db/db_connection')
+const app = require('../app')
+const mongoose = require('mongoose')
+
 const { filterData } = require('../api/shared/filter')
 
 describe('Filter out attributes from Query', () => {
@@ -22,3 +27,44 @@ describe('Filter out attributes from Query', () => {
     expect(options).toEqual({ sort: { timestamp: [-1] } })
   })
 })
+
+const testUser = {
+  name: 'test',
+  password: process.env.PASSWORD,
+  email: 'test3@gmail.com',
+  role: 'user',
+}
+
+const setupLogTests = async () => {
+  let cookieSession
+
+  await connectDB() // connect to local_db
+  // register user
+  await request(app).post('/api/user/register').send({
+    name: testUser.name,
+    password: testUser.password,
+    email: testUser.email,
+    role: testUser.role,
+  })
+  // login user and store cookie
+  await request(app)
+    .post('/api/user/login')
+    .send({
+      email: testUser.email,
+      password: testUser.password,
+    })
+    .then((res) => {
+      cookieSession = res.headers['set-cookie'][0]
+        .split(',')
+        .map((item) => item.split(';')[0])
+        .join(';')
+    })
+
+  return cookieSession
+}
+
+const teardownLogTests = async () => {
+  await mongoose.connection.close()
+}
+
+module.exports = { setupLogTests, teardownLogTests }
