+++ server/__tests__/disk_logs.test.js
@@ -1,55 +1,16 @@
 const request = require('supertest')
-const connectDB = require('../db/db_connection')
 const app = require('../app')
 const disk = require('../models/disk.js')
-const mongoose = require('mongoose')
-
-const testUser = {
-  name: 'test',
-  password: process.env.PASSWORD,
-  email: 'test3@gmail.com',
-  role: 'user',
-}
+const { setupLogTests, teardownLogTests } = require('./api_common.test')
 
 let cookieSession = ''
 
 beforeAll(async () => {
-  await connectDB() // connect to local_db
+  cookieSession = await setupLogTests()
   await disk.DiskLogs.deleteMany() //clear logs
-  // register user
-  await request(app).post('/api/user/register').send({
-    name: testUser.name,
-    password: testUser.password,
-    email: testUser.email,
-    role: testUser.role,
-  })
-  // login user and store cookie
-  await request(app)
-    .post('/api/user/login')
-    .send({
-      email: testUser.email,
-      password: testUser.password,
-    })
-    .then((res) => {
-      cookieSession = res.headers['set-cookie'][0]
-        .split(',')
-        .map((item) => item.split(';')[0])
-        .join(';')
-    })
 })
 
-const mockPayload = {
-  deviceId: 'B3C2D-C033-7B87-4B31-244BFE931F1E',
-  timestamp: '2021-10-24 09:47:55.966088',
-  processes: [
-    { name: 'python', pid: 12345 },
-    { name: 'celebid', pid: 12344 },
-  ],
-}
-
 describe('Check disk Logs from DB with DeviceID', () => {
-  const conditionalCPULogTest = (bool) => (bool ? test : test.skip)
-
   it('Should save the contents of a post to the DB', async () => {
     const response = await request(app)
       .get(
@@ -59,7 +20,7 @@ describe('Check disk Logs from DB with DeviceID', () => {
     expect(response.statusCode).toBe(200)
   })
 
-  it('Should save the contents of a post to the DB', async () => {
+  it('Should return error 500', async () => {
     const response = await request(app)
       .get('/api/disk-logs/specific-device?limit=2')
       .set('Cookie', cookieSession)
@@ -68,8 +29,6 @@ describe('Check disk Logs from DB with DeviceID', () => {
 })
 
 describe('Check disk Logs from DB with timestamps', () => {
-  const conditionalCPULogTest = (bool) => (bool ? test : test.skip)
-
   it('Should retrieve the contents of a post to the DB for a specific timestamp', async () => {
     const response = await request(app)
       .get(
@@ -99,8 +58,6 @@ describe('Check disk Logs from DB with timestamps', () => {
 })
 
 describe('Check disk Logs from DB with specific attributes', () => {
-  const conditionalCPULogTest = (bool) => (bool ? test : test.skip)
-
   it('Should retrieve the contents of a post to the DB for a specific timestamp and a deviceID', async () => {
     const response = await request(app)
       .get(
@@ -111,8 +68,6 @@ describe('Check disk Logs from DB with specific attributes', () => {
   })
 })
 
-afterAll((done) => {
-  // Closing the DB connection allows Jest to exit successfully.
-  mongoose.connection.close()
-  done()
+afterAll(async () => {
+  await teardownLogTests()
 })
