+++ server/__tests__/user.test.js
@@ -1,17 +1,9 @@
 const request = require('supertest')
-const connectDB = require('../db/db_connection')
 const User = require('../models/user.js')
 const app = require('../app')
-const mongoose = require('mongoose')
+const { setupLogTests, teardownLogTests } = require('./api_common.test')
 
 let cookieSession = ''
-
-const testUser = {
-  name: 'test',
-  password: process.env.PASSWORD,
-  email: 'test0@gmail.com',
-  role: 'user',
-}
 
 const userOne = {
   name: 'test',
@@ -21,28 +13,8 @@ const userOne = {
 }
 
 beforeAll(async () => {
-  await connectDB() // connect to local_db
+  cookieSession = await setupLogTests()
   await User.deleteMany()
-  // register user
-  await request(app).post('/api/user/register').send({
-    name: testUser.name,
-    password: testUser.password,
-    email: testUser.email,
-    role: testUser.role,
-  })
-  // login user and store cookie
-  await request(app)
-    .post('/api/user/login')
-    .send({
-      email: testUser.email,
-      password: testUser.password,
-    })
-    .then((res) => {
-      cookieSession = res.headers['set-cookie'][0]
-        .split(',')
-        .map((item) => item.split(';')[0])
-        .join(';')
-    })
 })
 
 describe('Sign up given a username and password', () => {
@@ -142,8 +114,6 @@ describe('Log out', () => {
   })
 })
 
-afterAll((done) => {
-  // Closing the DB connection allows Jest to exit successfully.
-  mongoose.connection.close()
-  done()
+afterAll(async () => {
+  await teardownLogTests()
 })
