+++ server/api/daemon.js
@@ -11,8 +11,7 @@ router.post('/', async (req, res) => {
     await newCpuLog.save()
     res.status(200).send()
   } catch (err) {
-    console.error(err.message)
-    res.status(500).send('Server Error')
+    res.status(500).send('Server Error ' + err.message)
   }
 })
