+++ server/api/wifiLogs.js
@@ -23,7 +23,6 @@ router.get('/specific-device', auth, async (req, res) => {
         return res.status(200).json(wifiLogs)
       })
   } catch (err) {
-    console.error(err.message)
     res.status(500).send('Server Error ' + err.message)
   }
 })
@@ -59,7 +58,6 @@ router.get('/timestamp', auth, async (req, res) => {
       })
     }
   } catch (err) {
-    console.error(err.message)
     res.status(500).send('Server Error ' + err.message)
   }
 })
