+++ server/db/db_connection.js
@@ -2,18 +2,14 @@ const mongoose = require('mongoose')
 const { exitProcess } = require('../destroyProcess')
 
 // connect to database
-const connectDB = async (printLog) => {
+const connectDB = async () => {
   try {
     await mongoose
       .connect(process.env.MONGODB_URL, {
         useNewUrlParser: true,
         useUnifiedTopology: true,
       })
-      .then(() => {
-        if (printLog) {
-          console.log('MongoDB connected...')
-        }
-      })
+      .then(() => console.log('MongoDB connected...'))
   } catch (err) {
     exitProcess(err.message, 1)
   }
